(lead-gen,research-thera): Root cause found. Two bugs conspired to produce "SUCCEEDED with Count 0":
**Primary bug (fixed)** — `backend/research_agent/books_graph.py:419` had a 14-placeholder `VALUES (...)` clause for 13 columns, so every book INSERT raised `psycopg.ProgrammingError: the query has 14 placeholders but 13 parameters were passed`. The persist node caught the error and returned `{"error": "persist failed: ..."}`.

**Secondary bug (already fixed in HEAD, but prod isn't on it)** — the old `generationJob` resolver (commit `655c949f`, still running in prod) ignored `state.error` and marked the job SUCCEEDED whenever LangGraph's run-level status was `"success"`. The persist error propagated into `state.message` and appeared as `result.message` while `count` went to 0. HEAD's `generationJob.ts:43-58` now checks `state.error` / `state.success === false` and marks such runs FAILED.

The fix I made:

```diff
-"VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s) "
+"VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s) "
```

That's all that was needed in code — the placeholder count now matches the 13 columns and 13-value tuple. No new logic required.

Next step to see the fix live on `researchthera.com`:
1. Deploy the backend (`pnpm run deploy` in `apps/research-thera/backend`, per the saved reference — needs Docker Desktop).
2. Deploy the Next.js app to Vercel (so HEAD's improved `generationJob.ts` ships and future persist errors actually mark the job FAILED).
3. Retry "Recommend Books" on journal 30.

Want me to also clean up the stale `fca6d7c4-...` job row, or leave it?

+++ apps/lead-gen/crates/lead-papers/src/pg.rs
@@ -9,7 +9,7 @@ pub async fn connect(url: &str) -> Result<PgPool> {
 #[allow(clippy::too_many_arguments)]
 pub async fn promote_contact(
     pg: &PgPool,
-    contact_id: i64,
+    contact_id: i32,
     github_handle: Option<&str>,
     papers_json: &serde_json::Value,
     gh_match_score: f32,
@@ -43,7 +43,7 @@ pub async fn promote_contact(
 
 #[derive(Debug, Clone, sqlx::FromRow)]
 pub struct ContactSeed {
-    pub id: i64,
+    pub id: i32,
     pub first_name: String,
     pub last_name: String,
     pub email: Option<String>,
